Change code because get_system_time() return value was changed in BeDIS

+++ src/LBeacon.h
@@ -250,8 +250,8 @@ typedef struct ThreadStatus {
 typedef struct ScannedDevice {
 
     char scanned_mac_address[LENGTH_OF_MAC_ADDRESS];
-    long long initial_scanned_time;
-    long long final_scanned_time;
+    int initial_scanned_time;
+    int final_scanned_time;
     /* List entries for linking the struct to scanned_list and
        tracked_BR_object_list or to tracked_BLE_object_list, depending
        whether the device type is BR_EDR or BLE. */
